#1520 moving the diirt home setting to the activator.
