+++ core/diirt/diirt-plugins/org.csstudio.diirt.util.preferences/plugin.xml
@@ -10,11 +10,5 @@
             name="Diirt preferences">
       </page>
    </extension>
-   <extension
-         point="org.csstudio.utility.product.module">
-      <workbench
-            class="org.csstudio.diirt.util.preferences.DiirtStartup">
-      </workbench>
-   </extension>
 
 </plugin>
