+++ core/diirt/diirt-plugins/org.csstudio.diirt.util/META-INF/MANIFEST.MF
@@ -11,9 +11,12 @@ Import-Package: org.diirt.datasource;version="3.0.1",
  org.diirt.service;version="3.0.1",
  org.diirt.util.config;version="3.0.1",
  org.eclipse.core.runtime;version="3.4.0",
+ org.eclipse.ui,
  org.osgi.framework;version="1.8.0"
 Bundle-ActivationPolicy: lazy
 Bundle-Activator: org.csstudio.diirt.util.Activator
 Require-Bundle: org.eclipse.equinox.preferences;bundle-version="3.5.200",
- org.eclipse.osgi
+ org.eclipse.osgi,
+ org.csstudio.startup;bundle-version="3.2.1",
+ org.csstudio.utility.product
 Export-Package: org.csstudio.diirt.util
